remove the home component
