+++ src/components/App/App.js
@@ -1,41 +1,27 @@
 import './App.css';
-import {
-  ApolloClient,
-  InMemoryCache,
-  ApolloProvider,
-  HttpLink,
-  from,
-} from '@apollo/client';
-import { onError } from '@apollo/client/link/error';
-import Home from '../Home/Home';
 
-const errorLink = onError(({ graphqlErrors, networkError }) => {
-  if (graphqlErrors) {
-    graphqlErrors.map(({ message, location, path }) => {
-      console.error(`GraphQl error ${message}`);
-      return null;
-    });
-  }
-});
-
-const link = from([
-  errorLink,
-  new HttpLink({
-    uri: `https://api.everbase.co/graphql?apikey=${process.env.REACT_APP_EVERBASE_API_KEY}`,
-  }),
-]);
-
-const client = new ApolloClient({
-  cache: new InMemoryCache(),
-  link: link,
-});
+import { useLazyQuery } from '@apollo/client';
+import Searchbar from '../Searchbar/Searchbar';
+import { COUNTRY_BY_CODE } from '../../graphQl/Queries';
+import CountryInfo from '../CountryInfo/CountryInfo';
 
 function App() {
+  const [getCountryByCode, { data, loading }] = useLazyQuery(COUNTRY_BY_CODE);
+
+  const onSearchChange = (searchInput) => {
+    // if the search is less then 2 characters, we can skip the query to the server
+    if (searchInput.length < 2) return;
+
+    getCountryByCode({
+      variables: { code: searchInput },
+    });
+  };
   return (
-    <ApolloProvider client={client}>
+    <div>
       <h1>GraphQL Country Client</h1>
-      <Home />
-    </ApolloProvider>
+      <Searchbar onChange={onSearchChange} />
+      <CountryInfo data={data} loading={loading} />
+    </div>
   );
 }
